function_app/requirements.txt
azure-functions==1.2.1
